# Conflicts:
#	README.md
#	pom.xml
#	sonar-project.properties
#	spring-cloud-gray-client-netflix/pom.xml
#	spring-cloud-gray-client-netflix/src/main/java/cn/springcloud/gray/client/netflix/RibbonServerChooser.java
#	spring-cloud-gray-client-netflix/src/main/java/cn/springcloud/gray/client/netflix/configuration/NetflixRibbonGrayAutoConfiguration.java
#	spring-cloud-gray-client-netflix/src/main/java/cn/springcloud/gray/client/netflix/ribbon/GrayChooserRule.java
#	spring-cloud-gray-client-netflix/src/main/java/cn/springcloud/gray/client/netflix/ribbon/GrayLoadBalanceRule.java
#	spring-cloud-gray-client-netflix/src/main/java/cn/springcloud/gray/client/netflix/ribbon/RibbonServerExplainer.java
#	spring-cloud-gray-client/pom.xml
#	spring-cloud-gray-client/src/main/java/cn/springcloud/gray/AbstractGrayManager.java
#	spring-cloud-gray-client/src/main/java/cn/springcloud/gray/GrayClientHolder.java
#	spring-cloud-gray-client/src/main/java/cn/springcloud/gray/GrayManager.java
#	spring-cloud-gray-client/src/main/java/cn/springcloud/gray/ServerChooser.java
#	spring-cloud-gray-client/src/main/java/cn/springcloud/gray/choose/DefaultGrayPredicate.java
#	spring-cloud-gray-client/src/main/java/cn/springcloud/gray/client/config/GrayClientDiscoveryAutoConfiguration.java
#	spring-cloud-gray-client/src/main/java/cn/springcloud/gray/decision/GrayDecisionDelegater.java
#	spring-cloud-gray-client/src/main/java/cn/springcloud/gray/request/track/GrayTrackHolder.java
#	spring-cloud-gray-client/src/main/java/cn/springcloud/gray/routing/connectionpoint/RoutingConnectionPoint.java
#	spring-cloud-gray-core/pom.xml
#	spring-cloud-gray-dependencies/pom.xml
#	spring-cloud-gray-plugins/pom.xml
#	spring-cloud-gray-plugins/spring-cloud-gray-plugin-eureka/pom.xml
#	spring-cloud-gray-plugins/spring-cloud-gray-plugin-eureka/src/main/java/cn/springcloud/gray/client/netflix/eureka/configuration/GrayClientEurekaAutoConfiguration.java
#	spring-cloud-gray-plugins/spring-cloud-gray-plugin-event-stream/pom.xml
#	spring-cloud-gray-plugins/spring-cloud-gray-plugin-feign/pom.xml
#	spring-cloud-gray-plugins/spring-cloud-gray-plugin-gateway/pom.xml
#	spring-cloud-gray-plugins/spring-cloud-gray-plugin-gateway/src/main/java/cn/springcloud/gray/client/gateway/GrayLoadBalancerClientFilter.java
#	spring-cloud-gray-plugins/spring-cloud-gray-plugin-gateway/src/main/java/cn/springcloud/gray/client/gateway/configuration/GrayGatewayAutoConfiguration.java
#	spring-cloud-gray-plugins/spring-cloud-gray-plugin-ribbon-nacos-discovery/pom.xml
#	spring-cloud-gray-plugins/spring-cloud-gray-plugin-ribbon-nacos-discovery/src/main/java/cn/springcloud/gray/client/plugin/ribbon/nacos/NacosServerExplainer.java
#	spring-cloud-gray-plugins/spring-cloud-gray-plugin-ribbon-nacos-discovery/src/main/java/cn/springcloud/gray/client/plugin/ribbon/nacos/configuration/GrayClientRibbonNacosAutoConfiguration.java
#	spring-cloud-gray-plugins/spring-cloud-gray-plugin-webflux/pom.xml
#	spring-cloud-gray-plugins/spring-cloud-gray-plugin-webmvc/pom.xml
#	spring-cloud-gray-plugins/spring-cloud-gray-plugin-zuul/pom.xml
#	spring-cloud-gray-samples/pom.xml
#	spring-cloud-gray-samples/spring-cloud-gray-btrace-sample/pom.xml
#	spring-cloud-gray-samples/spring-cloud-gray-eureka-sample/pom.xml
#	spring-cloud-gray-samples/spring-cloud-gray-gateway-sample/pom.xml
#	spring-cloud-gray-samples/spring-cloud-gray-gateway-sample/src/main/java/com/springcloud/gray/gateway/configuration/GatewayConfiguration.java
#	spring-cloud-gray-samples/spring-cloud-gray-gateway-sample/src/main/resources/config/application.yml
#	spring-cloud-gray-samples/spring-cloud-gray-performance-agent/pom.xml
#	spring-cloud-gray-samples/spring-cloud-gray-server-sample/pom.xml
#	spring-cloud-gray-samples/spring-cloud-gray-server-sample/src/main/resources/config/application.yml
#	spring-cloud-gray-samples/spring-cloud-gray-service-a-sample/pom.xml
#	spring-cloud-gray-samples/spring-cloud-gray-service-a1-sample/pom.xml
#	spring-cloud-gray-samples/spring-cloud-gray-service-b-sample/pom.xml
#	spring-cloud-gray-samples/spring-cloud-gray-service-b-sample/src/main/resources/config/application.yml
#	spring-cloud-gray-samples/spring-cloud-gray-stream-sample/pom.xml
#	spring-cloud-gray-samples/spring-cloud-gray-zuul-sample/pom.xml
#	spring-cloud-gray-server-plugins/pom.xml
#	spring-cloud-gray-server-plugins/spring-cloud-gray-server-plugin-eureka/pom.xml
#	spring-cloud-gray-server-plugins/spring-cloud-gray-server-plugin-event-stream/pom.xml
#	spring-cloud-gray-server-plugins/spring-cloud-gray-server-plugin-nacos-discovery/pom.xml
#	spring-cloud-gray-server/pom.xml
#	spring-cloud-gray-server/src/main/java/cn/springcloud/gray/server/dao/model/GrayServiceDO.java
#	spring-cloud-gray-server/src/main/java/cn/springcloud/gray/server/module/audit/ResultfulOpRecordAspect.java
#	spring-cloud-gray-server/src/main/java/cn/springcloud/gray/server/module/gray/GrayServerModule.java
#	spring-cloud-gray-server/src/main/java/cn/springcloud/gray/server/module/gray/jpa/JPAGrayServerModule.java
#	spring-cloud-gray-server/src/main/java/cn/springcloud/gray/server/resources/ExceptionTranslator.java
#	spring-cloud-gray-server/src/main/java/cn/springcloud/gray/server/resources/rest/GrayDecisionResource.java
#	spring-cloud-gray-server/src/main/java/cn/springcloud/gray/server/resources/rest/GrayInstanceResource.java
#	spring-cloud-gray-server/src/main/java/cn/springcloud/gray/server/resources/rest/GrayPolicyResource.java
#	spring-cloud-gray-starter-dependencies/pom.xml
#	spring-cloud-gray-utils/pom.xml
#	spring-cloud-gray-webui/src/views/gray-instance/complex-table.vue
#	spring-cloud-gray-webui/src/views/gray-policy/list.vue
#	spring-cloud-gray-webui/src/views/gray-trackor/complex-table.vue
#	spring-cloud-starter-gray-client/pom.xml
#	spring-cloud-starter-gray-client/src/main/resources/META-INF/spring.factories
#	spring-cloud-starter-gray-eureka-server/pom.xml
#	spring-cloud-starter-gray-server/pom.xml
#	spring-cloud-starter-gray-server/src/main/resources/META-INF/spring.factories

+++ spring-cloud-gray-server/src/main/java/cn/springcloud/gray/server/module/gray/jpa/JPAGrayServerModule.java
@@ -1,65 +1,51 @@
 package cn.springcloud.gray.server.module.gray.jpa;
 
-import cn.springcloud.gray.event.EventType;
-import cn.springcloud.gray.event.GrayEventMsg;
-import cn.springcloud.gray.event.GraySourceEventPublisher;
-import cn.springcloud.gray.event.SourceType;
 import cn.springcloud.gray.model.GrayStatus;
 import cn.springcloud.gray.model.InstanceInfo;
 import cn.springcloud.gray.model.InstanceStatus;
 import cn.springcloud.gray.server.configuration.properties.GrayServerProperties;
 import cn.springcloud.gray.server.discovery.ServiceDiscovery;
 import cn.springcloud.gray.server.module.gray.GrayServerModule;
-import cn.springcloud.gray.server.module.gray.domain.GrayDecision;
 import cn.springcloud.gray.server.module.gray.domain.GrayInstance;
-import cn.springcloud.gray.server.module.gray.domain.GrayPolicy;
 import cn.springcloud.gray.server.module.gray.domain.GrayService;
+import cn.springcloud.gray.server.module.gray.domain.query.GrayServiceQuery;
 import cn.springcloud.gray.server.module.user.ServiceManageModule;
-import cn.springcloud.gray.server.service.GrayDecisionService;
 import cn.springcloud.gray.server.service.GrayInstanceService;
-import cn.springcloud.gray.server.service.GrayPolicyService;
 import cn.springcloud.gray.server.service.GrayServiceService;
+import cn.springlcoud.gray.event.server.GrayEventTrigger;
+import cn.springlcoud.gray.event.server.TriggerType;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.Arrays;
 import java.util.Collection;
+import java.util.Iterator;
 import java.util.List;
 import java.util.Objects;
 
 @Slf4j
 public class JPAGrayServerModule implements GrayServerModule {
 
-    private static final List<SourceType> INSTANCE_STATUS_CHECK_SOURCE_TYPES =
-            Arrays.asList(SourceType.GRAY_POLICY, SourceType.GRAY_DECISION);
-
     private GrayServiceService grayServiceService;
     private GrayInstanceService grayInstanceService;
-    private GrayDecisionService grayDecisionService;
-    private GrayPolicyService grayPolicyService;
-    private GraySourceEventPublisher grayEventPublisher;
+    private GrayEventTrigger grayEventTrigger;
     private GrayServerProperties grayServerProperties;
     private ServiceDiscovery serviceDiscovery;
     private ServiceManageModule serviceManageModule;
 
     public JPAGrayServerModule(
             GrayServerProperties grayServerProperties,
-            GraySourceEventPublisher grayEventPublisher,
+            GrayEventTrigger grayEventTrigger,
             ServiceDiscovery serviceDiscovery,
             GrayServiceService grayServiceService,
             GrayInstanceService grayInstanceService,
-            GrayDecisionService grayDecisionService,
-            GrayPolicyService grayPolicyService,
             ServiceManageModule serviceManageModule) {
         this.grayServerProperties = grayServerProperties;
-        this.grayEventPublisher = grayEventPublisher;
+        this.grayEventTrigger = grayEventTrigger;
         this.serviceDiscovery = serviceDiscovery;
         this.grayServiceService = grayServiceService;
         this.grayInstanceService = grayInstanceService;
-        this.grayDecisionService = grayDecisionService;
-        this.grayPolicyService = grayPolicyService;
         this.serviceManageModule = serviceManageModule;
     }
 
@@ -79,16 +65,24 @@ public class JPAGrayServerModule implements GrayServerModule {
         return grayInstanceService.listGrayInstancesByNormalInstanceStatus(instanceStatus);
     }
 
+    @Override
+    public List<GrayInstance> listGrayInstances(Iterator<String> serviceIds, Collection<InstanceStatus> instanceStatus) {
+        return grayInstanceService.listGrayInstances(serviceIds, instanceStatus);
+    }
+
 
     @Transactional
     @Override
     public void deleteGrayService(String serviceId) {
-        List<GrayInstance> grayInstances = grayInstanceService.findByServiceId(serviceId);
+        GrayService grayService = getGrayService(serviceId);
+        if (Objects.isNull(grayService)) {
+            return;
+        }
         grayServiceService.deleteReactById(serviceId);
-        serviceManageModule.deleteSericeManeges(serviceId);
+//        List<GrayInstance> grayInstances = grayInstanceService.findByServiceId(serviceId);
 //        grayInstances.forEach(this::publishDownIntanceEvent);
-        publishEventMsg(
-                createEventMsg(SourceType.GRAY_SERVICE, EventType.DOWN, serviceId), null);
+        serviceManageModule.deleteSericeManeges(serviceId);
+        triggerDeleteEvent(grayService);
     }
 
 
@@ -99,9 +93,9 @@ public class JPAGrayServerModule implements GrayServerModule {
             instance.setGrayStatus(grayStatus);
             grayInstanceService.saveModel(instance);
             if (grayStatus == GrayStatus.OPEN) {
-                publishUpdateIntanceEvent(instance);
+                triggerUpdateEvent(instance);
             } else {
-                publishDownIntanceEvent(instance);
+                triggerDeleteEvent(instance);
             }
         }
     }
@@ -132,13 +126,14 @@ public class JPAGrayServerModule implements GrayServerModule {
         if (isLockGray(newRecord) ||
                 grayServerProperties.getInstance().getNormalInstanceStatus().contains(instance.getInstanceStatus())) {
             if (Objects.equals(newRecord.getGrayStatus(), GrayStatus.OPEN)) {
-                publishUpdateIntanceEvent(newRecord);
+                triggerUpdateEvent(instance);
             } else if (oldRecord != null && Objects.equals(oldRecord.getGrayStatus(), GrayStatus.OPEN)) {
-                publishDownIntanceEvent(newRecord);
+                triggerDeleteEvent(instance);
             }
         }
         return newRecord;
     }
+
 
     @Override
     public void updateInstanceStatus(String instanceId, InstanceStatus instanceStatus) {
@@ -157,9 +152,9 @@ public class JPAGrayServerModule implements GrayServerModule {
             }
             if (instance.getGrayStatus() == GrayStatus.OPEN) {
                 if (grayServerProperties.getInstance().getNormalInstanceStatus().contains(instanceStatus)) {
-                    publishUpdateIntanceEvent(instance);
+                    triggerUpdateEvent(instance);
                 } else {
-                    publishDownIntanceEvent(instance);
+                    triggerDeleteEvent(instance);
                 }
             }
         }
@@ -171,7 +166,6 @@ public class JPAGrayServerModule implements GrayServerModule {
         GrayInstance grayInstance = grayInstanceService.findOneModel(intanceId);
         if (grayInstance != null) {
             grayInstanceService.deleteReactById(intanceId);
-            publishDownIntanceEvent(grayInstance);
         }
     }
 
@@ -179,65 +173,6 @@ public class JPAGrayServerModule implements GrayServerModule {
         return Objects.equals(instance.getGrayLock(), GrayInstance.GRAY_LOCKED);
     }
 
-    @Override
-    public GrayPolicy saveGrayPolicy(GrayPolicy grayPolicy) {
-        GrayPolicy newRecord = grayPolicyService.saveModel(grayPolicy);
-//        publishUpdateIntanceEvent(grayInstanceService.findOneModel(grayPolicy.getInstanceId()));
-        publishEventMsg(
-                createEventMsg(SourceType.GRAY_POLICY, EventType.UPDATE,
-                        grayInstanceService.findOneModel(newRecord.getInstanceId())), newRecord);
-        return newRecord;
-    }
-
-    @Transactional
-    @Override
-    public void deleteGrayPolicy(Long policyId) {
-        GrayPolicy policy = grayPolicyService.findOneModel(policyId);
-        if (policy != null) {
-            grayPolicyService.deleteReactById(policyId);
-//            publishUpdateIntanceEvent(grayInstanceService.findOneModel(policy.getInstanceId()));
-            publishEventMsg(
-                    createEventMsg(SourceType.GRAY_POLICY, EventType.DOWN,
-                            grayInstanceService.findOneModel(policy.getInstanceId())), policy);
-        }
-    }
-
-    @Override
-    public GrayDecision saveGrayDecision(GrayDecision grayDecision) {
-        GrayPolicy policy = grayPolicyService.findOneModel(grayDecision.getPolicyId());
-        grayDecision.setInstanceId(policy.getInstanceId());
-        GrayDecision newRecord = grayDecisionService.saveModel(grayDecision);
-//        publishUpdateIntanceEvent(grayInstanceService.findOneModel(grayDecision.getInstanceId()));
-
-        publishEventMsg(
-                createEventMsg(SourceType.GRAY_DECISION, EventType.UPDATE,
-                        grayInstanceService.findOneModel(newRecord.getInstanceId())), newRecord);
-        return newRecord;
-    }
-
-    @Override
-    public void deleteGrayDecision(Long decisionId) {
-        GrayDecision decision = grayDecisionService.findOneModel(decisionId);
-        if (decision != null) {
-            grayDecisionService.delete(decisionId);
-            publishEventMsg(
-                    createEventMsg(SourceType.GRAY_DECISION, EventType.DOWN,
-                            grayInstanceService.findOneModel(decision.getInstanceId())), decision);
-//            publishUpdateIntanceEvent(grayInstanceService.findOneModel(decision.getInstanceId()));
-        }
-
-    }
-
-
-    @Override
-    public GrayDecision getGrayDecision(Long id) {
-        return grayDecisionService.findOneModel(id);
-    }
-
-    @Override
-    public List<GrayDecision> listGrayDecisionsByPolicyId(Long policyId) {
-        return grayDecisionService.findByPolicyId(policyId);
-    }
 
     @Override
     public List<GrayInstance> listGrayInstancesByServiceId(String serviceId) {
@@ -266,8 +201,7 @@ public class JPAGrayServerModule implements GrayServerModule {
         if (serviceManageModule.getServiceOwner(grayService.getServiceId()) == null) {
             serviceManageModule.addServiceOwner(grayService.getServiceId());
         }
-        publishEventMsg(
-                createEventMsg(SourceType.GRAY_SERVICE, EventType.UPDATE, record.getServiceId()), record);
+        triggerUpdateEvent(record);
         return record;
     }
 
@@ -277,13 +211,13 @@ public class JPAGrayServerModule implements GrayServerModule {
     }
 
     @Override
-    public List<GrayPolicy> listGrayPoliciesByInstanceId(String instanceId) {
-        return grayPolicyService.findByInstanceId(instanceId);
+    public Page<GrayService> listAllGrayServices(Pageable pageable) {
+        return grayServiceService.listAllGrayServices(pageable);
     }
 
     @Override
-    public Page<GrayService> listAllGrayServices(Pageable pageable) {
-        return grayServiceService.listAllGrayServices(pageable);
+    public Page<GrayService> queryGrayServices(GrayServiceQuery query, Pageable pageable) {
+        return grayServiceService.queryGrayServices(query, pageable);
     }
 
     @Override
@@ -292,76 +226,47 @@ public class JPAGrayServerModule implements GrayServerModule {
     }
 
     @Override
-    public Page<GrayPolicy> listGrayPoliciesByInstanceId(String instanceId, Pageable pageable) {
-        return grayPolicyService.listGrayPoliciesByInstanceId(instanceId, pageable);
-    }
-
-    @Override
     public Page<GrayInstance> listGrayInstancesByServiceId(String serviceId, Pageable pageable) {
         return grayInstanceService.listGrayInstancesByServiceId(serviceId, pageable);
     }
 
-    @Override
-    public Page<GrayDecision> listGrayDecisionsByPolicyId(Long policyId, Pageable pageable) {
-        return grayDecisionService.listGrayDecisionsByPolicyId(policyId, pageable);
-    }
-
-    protected GraySourceEventPublisher getGrayEventPublisher() {
-        return grayEventPublisher;
+    protected GrayEventTrigger getGrayEventTrigger() {
+        return grayEventTrigger;
     }
 
 
-    private void publishUpdateIntanceEvent(GrayInstance grayInstance) {
-        GrayEventMsg eventMsg = createEventMsg(SourceType.GRAY_INSTANCE, EventType.UPDATE, grayInstance);
-        if (eventMsg != null) {
-            log.info("push event message -> {}", eventMsg);
-            getGrayEventPublisher().publishEvent(eventMsg, grayInstance, 100);
-        }
+    protected void triggerEvent(TriggerType triggerType, Object source) {
+        getGrayEventTrigger().triggering(source, triggerType);
     }
 
-
-    private GrayEventMsg createEventMsg(SourceType sourceType, EventType eventType, GrayInstance grayInstance) {
-        if (!isNeesPushEentMsg(sourceType, grayInstance)) {
-            return null;
-        }
-        return GrayEventMsg.builder()
-                .serviceId(grayInstance.getServiceId())
-                .instanceId(grayInstance.getInstanceId())
-                .eventType(eventType)
-                .sourceType(sourceType)
-                .build();
+    protected void triggerDeleteEvent(Object source) {
+        triggerEvent(TriggerType.DELETE, source);
     }
 
-    private boolean isNeesPushEentMsg(SourceType sourceType, GrayInstance grayInstance) {
-        if (INSTANCE_STATUS_CHECK_SOURCE_TYPES.contains(sourceType)) {
-            return Objects.equals(grayInstance.getGrayStatus(), GrayStatus.OPEN)
-                    && (
-                    isLockGray(grayInstance) ||
-                            grayServerProperties.getInstance().getNormalInstanceStatus().contains(grayInstance.getInstanceStatus()));
-        }
-        return true;
-    }
-
-    private GrayEventMsg createEventMsg(SourceType sourceType, EventType eventType, String serviceId) {
-        return GrayEventMsg.builder()
-                .serviceId(serviceId)
-                .eventType(eventType)
-                .sourceType(sourceType)
-                .build();
-    }
-
-
-    private void publishDownIntanceEvent(GrayInstance grayInstance) {
-        publishEventMsg(
-                createEventMsg(SourceType.GRAY_INSTANCE, EventType.DOWN, grayInstance), grayInstance);
-    }
-
-    public void publishEventMsg(GrayEventMsg eventMsg, Object source) {
-        if (eventMsg == null) {
+    protected void triggerUpdateEvent(Object source) {
+        if (source instanceof GrayInstance && !isActiveGrayInstance((GrayInstance) source)) {
+            GrayInstance grayInstance = (GrayInstance) source;
+            log.info("服务{} 实例{}非灰度状态，不同步事件信息。", grayInstance.getServiceId(), grayInstance.getInstanceId());
             return;
         }
-        log.info("push event message -> {}", eventMsg);
-        getGrayEventPublisher().asyncPublishEvent(eventMsg, source);
+        triggerEvent(TriggerType.MODIFY, source);
     }
 
+
+    @Override
+    public boolean isActiveGrayInstance(String instanceId) {
+        GrayInstance grayInstance = getGrayInstance(instanceId);
+        if (Objects.isNull(grayInstance)) {
+            return false;
+        }
+        return isActiveGrayInstance(grayInstance);
+    }
+
+    @Override
+    public boolean isActiveGrayInstance(GrayInstance grayInstance) {
+        return Objects.equals(grayInstance.getGrayStatus(), GrayStatus.OPEN)
+                && (
+                isLockGray(grayInstance) ||
+                        grayServerProperties.getInstance().getNormalInstanceStatus().contains(grayInstance.getInstanceStatus()));
+    }
 }
